Initial Commit
This commit lays out the basic project structure, and includes all of the code
provided with the sample project that help with drawing a window and handling
user input.

In addition, the solution to the first homework problem is included. To compile
the binary type 'make part1' into the console and then execute the 'bin/part1'
file.